remove ChannelInner component
