+++ src/components/Channel/index.js
@@ -1,2 +1,1 @@
 export { default as Channel } from './Channel';
-export { default as ChannelInner } from './ChannelInner';
